+++ server/package.json
@@ -11,7 +11,7 @@
     "start": "node -r dotenv/config dist/index.js",
     "start:dev": "concurrently \"tsc -w\" \"nodemon dist/index.js\"",
     "test:all": "jest --maxWorkers =1",
-    "test:unit": "jest --testPathPattern=unit.ts$ --maxWorkers=1",
+    "test:unit": "jest --testPathPattern=unit.ts$ --runInBand",
     "test:integration": "jest --testPathPattern=integration.ts$ --runInBand"
   },
   "dependencies": {
