+++ server/src/__tests__/club.test.integration.ts
@@ -8,13 +8,9 @@ app.use(express.json())
 app.use('/api/v1', v1Router)
 
 describe('Club Router', () => {
-  beforeAll(async () => {
-    await populate()
-  })
+  beforeAll(populate)
 
-  afterAll(async () => {
-    await teardown()
-  })
+  afterAll(teardown)
 
   test('When a GET req is fired to /clubs, it should return all clubs', async () => {
     const res = await request(app).get('/api/v1/clubs').set('Accept', 'application/json')
