+++ server/src/modules/events/mappers/event-dto.ts
@@ -1,0 +1,16 @@
+import { BasicClub } from "../domain/entities/event";
+
+export interface EventDTO {
+  name: string
+  description: string
+  url?: string
+  startTime: string
+  endTime: string
+  facebookLink?: string
+  twitterLink?: string
+  instagramLink?: string
+  websiteLink?: string
+  backgroundImageURL?: string
+  tags: Array<string>
+  clubs: Array<BasicClub>
+}
